Added layout to load navbar and sidebar
