+++ riddler/src/components/layout/navbar/assets/riddler-logo.png
